+++ ceasiompy/EdgeRun/func/edgeutils.py
@@ -23,12 +23,6 @@ from pathlib import Path
 
 import requests
 from ceasiompy.utils.ceasiomlogger import get_logger
-from ceasiompy.utils.ceasiompyutils import get_install_path
-from ceasiompy.utils.commonnames import (
-    ACTUATOR_DISK_INLET_SUFFIX,
-    ACTUATOR_DISK_OUTLET_SUFFIX,
-    ENGINE_EXHAUST_SUFFIX,
-    ENGINE_INTAKE_SUFFIX,
 )
 from ceasiompy.utils.moduleinterfaces import get_module_path
 
@@ -57,7 +51,7 @@ def get_edge_ainp_template():
         )
     return edge_ainp_template_path
 
-def get_edge_queScript_template():
+def get_edge_que_script_template():
     """Return path of the M-Edge ainp template corresponding to the M-Edge version."""
 
     edge_dir = get_module_path("EdgeRun")
